#实现了简单的调度器以及 controller 增加了添加 dockerd 格式校验
docker地址格式

* http://192.168.119.10:2375
* https://192.168.119.10:2375
* unix:///var/run/docker.sock

+++ models/unit.go
@@ -4,7 +4,7 @@ type Unit struct {
 	Id          int64
 	Name        string `orm:"unique"`
 	Image       string
-	Number      uint
+	Number      int64
 	Parameteres []*UnitParameter `orm:"reverse(many);on_delete(set_null)"`
 	Dockerd     []*Dockerd       `orm:"rel(m2m)"`
 }
